Safari-611.1.10/LayoutTests/inspector/canvas/requestShaderSource-expected.txt

Common tests for Canvas.requestShaderSource command.


== Running test suite: Canvas.requestShaderSource
-- Running test case: Canvas.requestShaderSource.ProgramId.Invalid
PASS: Should produce an error.
Error: Unknown shaderType: INVALID_SHADER_TYPE
